# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "api-server",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/admin/sync/start": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Start sync for all enabled mailing lists",
        "operationId": "routes_admin_start_sync",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/sync/queue": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Start sync for specific mailing lists",
        "operationId": "routes_admin_queue_sync",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/sync/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get current sync status (running + queued jobs)",
        "operationId": "routes_admin_get_sync_status",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/sync/cancel": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Cancel ALL sync jobs including currently running ones",
        "operationId": "routes_admin_cancel_sync",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/database/reset": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Reset the database (drop and recreate all tables)",
        "operationId": "routes_admin_reset_db",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/database/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get database statistics (global, not per mailing list) Queries are parallelized for faster response times",
        "operationId": "routes_admin_get_database_status",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
    "/admin/database/config": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get PostgreSQL configuration for monitoring",
        "operationId": "routes_admin_get_database_config",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "SyncStartResponse": {
        "type": "object",
        "required": [
          "jobIds",
          "message"
        ],
        "properties": {
          "jobIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "SyncRequest": {
        "type": "object",
        "required": [
          "mailingListSlugs"
        ],
        "properties": {
          "mailingListSlugs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SyncStatusResponse": {
        "type": "object",
        "required": [
          "isRunning",
          "queuedJobs"
        ],
        "properties": {
          "currentJob": {
            "allOf": [
              {
                "$ref": "#/components/schemas/JobStatusInfo"
              }
            ],
            "nullable": true
          },
          "queuedJobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueuedJobInfo"
            }
          },
          "isRunning": {
            "type": "boolean"
          }
        }
      },
      "JobStatusInfo": {
        "type": "object",
        "required": [
          "created_at",
          "id",
          "mailing_list_id",
          "name",
          "phase",
          "priority",
          "slug"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mailing_list_id": {
            "type": "integer",
            "format": "int32"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "error_message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "QueuedJobInfo": {
        "type": "object",
        "required": [
          "id",
          "mailingListId",
          "mailingListName",
          "mailingListSlug",
          "position"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mailingListId": {
            "type": "integer",
            "format": "int32"
          },
          "mailingListSlug": {
            "type": "string"
          },
          "mailingListName": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DatabaseStatusResponse": {
        "type": "object",
        "required": [
          "totalAuthors",
          "totalEmails",
          "totalRecipients",
          "totalReferences",
          "totalThreadMemberships",
          "totalThreads"
        ],
        "properties": {
          "totalAuthors": {
            "type": "integer",
            "format": "int64"
          },
          "totalEmails": {
            "type": "integer",
            "format": "int64"
          },
          "totalThreads": {
            "type": "integer",
            "format": "int64"
          },
          "totalRecipients": {
            "type": "integer",
            "format": "int64"
          },
          "totalReferences": {
            "type": "integer",
            "format": "int64"
          },
          "totalThreadMemberships": {
            "type": "integer",
            "format": "int64"
          },
          "dateRangeStart": {
            "type": "string",
            "format": "partial-date-time",
            "nullable": true
          },
          "dateRangeEnd": {
            "type": "string",
            "format": "partial-date-time",
            "nullable": true
          }
        }
      }
    }
  }
}